+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * add TemplateEngine for general purpose
 * module: renamed bind(supplier) to bindSupplier()
 * template: "include" support
+* queue: renamed MessagePublisher.reply to publish
 
 ### 3.2.7 (7/21/2015 - 7/22/2015)
 * db repository, added selectAll()